tests/decode-incomplete-input.html

<!DOCTYPE html>
<script src="resources/text-encode-transform.js"></script>
<script src="resources/transform-stream.js"></script>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/readable-stream-from-array.js"></script>
<script src="resources/readable-stream-to-array.js"></script>
<script>
'use strict';

const inputBytes = [229];

promise_test(() => {
  const input = readableStreamFromArray([new Uint8Array(inputBytes)]);
  const output = input.pipeThrough(new TextDecoder());
  return readableStreamToArray(output)
      .then(array => {
        assert_array_equals(array, ['', '\uFFFD'], 'array should match expectation');
      });
}, 'incomplete input with error mode "replacement" should end with a ' +
    'replacement character');

promise_test(t => {
  const input = readableStreamFromArray([new Uint8Array(inputBytes)]);
  const output = input.pipeThrough(
      new TextDecoder('utf-8', {fatal: true}));
  const reader = output.getReader();
  return reader.read()
      .then(result => {
        assert_equals(result.value, '', 'first chunk should be empty');
        return promise_rejects(t, new TypeError(), reader.read(),
                               'second read should reject');
      });
}, 'incomplete input with error mode "fatal" should error the stream');

</script>
